src/requirements.txt
Flask
paho-mqtt
Pillow
pytz
